nu merge create LocationInLibrary

+++ BookLibrary/Models/LocationInLibraryModel.cs
@@ -9,8 +9,8 @@ namespace BookLibrary.Models
     {
         public Guid IDLocationInLibrary { get; set; }
         public string Name { get; set; }
-        public int Floor { get; set; }
-        public int Sector { get; set; }
+        public decimal Floor { get; set; }
+        public decimal Sector { get; set; }
         public string Shelf { get; set; }
     }
 }
